requirements.txt
fastapi==0.88.0
Pillow==9.5.0
protobuf==4.25.0
python-multipart==0.0.6
sentencepiece==0.1.99
transformers==4.35.0
transformers[torch]
uvicorn==0.20.0
